Integrate all roles together for full deployment
This commit brings all existing roles together and integrates them. It
makes it possible for someone to use these roles in conjunction with
each other, such as for a server configuration. It allows us to move
forward with running it on our infrastructure.

A summation of changes:

MariaDB
-------

* Copy defaults from current MariaDB configuration

php-fpm
-------

* Minor syntax cleanup

nginx
-----

* Split configurations out into three different globs
  * **restrictions.conf**: Base set of restrictions for all sites
  * **wordpress.conf**: Base set of guidelines for any WordPress site
  * **fossrit.jwf.io**: Single installed WordPress site
* Adjust various settings with PHP

Ansible
-------

* Update nginx role task to load in new global config files

Resources
---------

Resources informing this commit:

* https://codex.wordpress.org/Nginx
* https://www.nginx.com/resources/wiki/start/topics/recipes/wordpress/