[matterbridge] Remove {PROTOCOL} from RemoteNickFormat in IRC
This amends 0e3c14d232335699ffbe84b7504cebce7342458a, which actually
only changed the nick format, not the protocol prefix. :| Not sure how I
made that mistake. Anyways…

This actually closes #28.

And my vim settings trimmed all of the hanging whitespace originally in
this file. That's why the diff is so huge…