docs: Document how fossrit.community Discourse website is deployed
Closes FOSSRIT/tasks#90.

This commit adds a new document written in AsciiDoc to explain how the
Discourse website infrastructure is deployed. This intends to create a
trail of how the application is deployed and how changes can be made
(and who to contact) in the future.

This needs more expansion, like how DNS is handled for the
fossrit.community domain, but this is a start for now.