matterbridge: Update to v1.15.0, update distribution URLs
This commit changes the following:

* Updates Matterbridge to latest upstream version, v1.15.0
* Changes distribution method to follow latest format by upstream

The v1.15.0 upgrade includes a bugfix from v1.14.2 that affected the IRC
bridge. In previous versions, the bot would not reconnect automatically
or gracefully if it lost connection to the IRC server. Now, the bot
should automatically reconnect if the connection is terminated or broken
for some reason. cc: @Tjzabel @ct-martin

Additionally, upstream also ships releases in a new format. I changed
the URL schema to match the new format and changed the checksums to
sha256sum since upstream provides those in a release. Previously, I was
calculating sha512sums manually, which isn't exactly straightforward.