:slot_machine: Add metrics host to inventory (#62)
This commit adds the College of Art and Design virtual machine in
Institute Hall to the Ansible inventory. Along with it, I added a new
base role for an Ubuntu 18.04 LTS machine, since this is the operating
system that the CAD admins use as a base. Additionally, I also added a
new playbook to run the role on the host. I tested this locally before
committing and verified it all works.

Currently, I'm hard-coding my RIT computer account as the login user,
which obviously is a hard block on me. A future next step is to set up
an Ansible system user for running playbooks. I need to think about this
harder before implementing, but in the meanwhile, this commit gets us
rolling.

cc: @Nolski @Tjzabel

Signed-off-by: Justin W. Flory <git@jwf.io>