:snake: teleirc: Add bridge for #rit-python / @rit_python
This bridges the IRC channel (ircs://chat.freenode.net:6697/#rit-python)
with the Telegram group (https://t.me/rit_python). This is the beginning
of a small social experiment to bridge a few chats together.

If this is successful, we may try bridging to other RIT community chat
rooms and see what happens.

Signed-off-by: Justin W. Flory <git@jwf.io>