:wastebasket: Migrate from Vagrant to Pipfile-driven CI
This isn't a true migration since the purpose of Vagrant, to fully test
the Ansible playbooks, is not being leveraged. Since they are not being
used or maintained, this commit swaps the Vagrantfile out with a
Pipfile. This guarantees a fixed version of `ansible-playbook` when
running inside of a `pipenv shell` environment. It doesn't guarantee
that the playbooks actually work though.

Eventually, I will circle back to this to see if we can integrate a more
thorough CI, perhaps with Vagrant, or perhaps with containers. TBD.

Signed-off-by: Justin W. Flory <git@jwf.io>