:bridge_at_night: matterbridge: Unify IRC channels to RIT namespace (#68)
This commit reorganizes the channels used for the `#ritlug`/`#general`
channel and the `#ritlug-teleirc`/`#teleirc` channels. This change is
done because RIT/FOSS@MAGIC is registered to the `#rit-*` namespace. Any
channels that fall underneath the umbrella of the `#rit-*` namespace
enable Freenode operators to provide support to organization admins to
troubleshoot or change permissions in channels. This also allows us to
follow a better categorization system for channels, e.g. FOSS@MAGIC
admins control `#rit-*` channels, but RITlug eboard controls
`#rit-lug-*` channels. Just an example!

The commit does exactly the following:

1. Renames vars file channels to use RIT namespace format
2. Adds temporary manual entry in matterbridge config to route traffic
   into `#ritlug` and `#ritlug-teleirc`, before officially deprecating

I'm not sure how #2 will work but it will be a cool experiment to see
what happens, since we have RITlug/teleirc as a middleware of sorts
here.

CC: @Nolski @Tjzabel

Signed-off-by: Justin W. Flory <git@jwf.io>