:pencil2: matterbridge: Deprecate #ritlug/#ritlug-teleirc IRC channels (#72)
This commit retires the deprecated channels by removing the Matterbridge
bot from the old channels. Last week, we successfully migrated the IRC
channels to the RIT-registered namespace on Freenode. The two-way bridge
across two channels will be removed once the playbook runs with this
commit included.

Also, the diff is hard to read because I reorganized the channels
alphabetically to make it easier to maintain this file if it grows over
time. Sorry.

Signed-off-by: Justin W. Flory <git@jwf.io>