:arrow_up: matterbridge: Upgrade to v1.17.0 release
Matterbridge upstream released a new minor feature release:

    https://github.com/42wim/matterbridge/releases/tag/v1.17.0

There are a good number of changes, including support for Microsoft
Teams. (What?!) But the relevant bits of the changelog for us are below:

* general: support JSON and YAML config formats
  (42wim/matterbridge#1045)
* irc: Be less lossy when throttling IRC messages
  (42wim/matterbridge#1004)
* slack: Use upstream slack-go/slack again (42wim/matterbridge#1018)
* slack: Ignore ConnectingEvent (42wim/matterbridge#1041)
* slack: use blocks not attachments (42wim/matterbridge#1048)
* slack: Fix 42wim/matterbridge#1039: messages sent to Slack being
  synced back (42wim/matterbridge#1046)

And a hat tip to our upstream devs who made this release possible:
@qaisjp, @jakubgs, @burner1024, @notpushkin, @MartijnBraam, and @42wim.

This upgrade is currently running in production without issue. This
commit brings the config management up to date with what is currently
running out in prod.

Signed-off-by: Justin W. Flory <git@jwf.io>