:bridge_at_night: teleirc: Add bridge for RIT group chat (closes #77)
This commit adds a new TeleIRC bridge for the Rochester Institute of
Technology group chat on Freenode and Telegram. This is `#rit` on
Freenode and https://t.me/rittigers on Telegram.

Thanks to @Tjzabel for creating the group and registering the bot, since
I've maxed out my registrations for both of these things now.
:see_no_evil:

Closes #77.

Signed-off-by: Justin W. Flory <git@jwf.io>